requirements.txt
django==3.2.8
djangorestframework==3.12.4
markdown==3.3.4
python-dotenv==0.19.1
dj-rest-auth==2.1.11
djangorestframework-simplejwt==5.0.0
psycopg2-binary==2.9.1
django-redis==5.0.0
pillow==8.4.0
django-filter==21.1